requirements.txt
cadquery>=2.5.2
trimesh>=4.4.0
matplotlib>=3.9.0
openai>=1.30.1
google-generativeai>=0.7.0
requests>=2.31.0
tinker @ git+https://github.com/thinking-machines-lab/tinker.git
